floating point accel/gyro, started creating analogRead replacement
- Accelerometer and gyroscope readings now passed as floats (way more
accurate)
- Working to create a replacement function for analogRead that preserves
the ADC ISR. I've got some weird error right now where I get repeat
values across ADCs, I believe it's due to some issue with the ADC MUX.
More to come...

+++ SeaSense_ArduinoLib/globals.h
@@ -68,8 +68,8 @@
     extern unsigned long Light;
     extern int carryOut;
     extern int Head;
-    extern int AccelX,AccelY,AccelZ;
-    extern int GyroX,GyroY,GyroZ;
+    extern float AccelX,AccelY,AccelZ;
+    extern float GyroX,GyroY,GyroZ;
     extern int adcBuf[ADC_BUFFER_SIZE];
     extern byte adc_channel, adc_pos;
     extern int vBat;
